requirements.txt
streamlit
yfinance
pandas
pandas-ta
requests
python-dotenv
plotly
schedule
sqlalchemy
